Исправил код в соответствии с рекомендациями
