+++ src/components/app/app.module.css
@@ -9,4 +9,5 @@
   display: flex;
   width: 80%;
   gap: 40px;
+  justify-content: center;
 }
