+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo } from "react";
+import { useMemo } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { useDrop } from "react-dnd";
 
@@ -11,16 +11,10 @@ import BurgerConstructorConfirm from "./burger-constructor-confirm";
 import FillingsContainer from "./fillings-container";
 
 // Utils
-import { countTotal, isEmptyConstuctor, checkBun } from "../../utils/helpers";
-import { getIngredientsId } from "../../utils/helpers";
-import { setOrderId } from "../../services/reducers/order";
+import { checkBun } from "../../utils/helpers";
 
 // ACTIONS-REDUCERS
-import {
-  COUNT_TOTAL,
-  ADD_FILLING,
-  ADD_BUN,
-} from "../../services/actions/constructor";
+import { ADD_FILLING, ADD_BUN } from "../../services/actions/constructor";
 import {
   INGREDIENTS_COUNTER_INCREASE,
   CHANGE_BUN,
@@ -30,12 +24,7 @@ function BurgerConstructor() {
   const { ingredients } = useSelector((store) => store.constructorBurger);
   const dispatch = useDispatch();
 
-  const isEmpty = useMemo(() => isEmptyConstuctor(ingredients), [ingredients]);
   const isEmptyBun = useMemo(() => checkBun(ingredients), [ingredients]);
-  const memoizedTotal = useMemo(
-    () => countTotal(ingredients, isEmpty),
-    [ingredients, isEmpty]
-  );
 
   const [, dropTarget] = useDrop({
     accept: ["bun", "main", "sauce"],
@@ -65,27 +54,6 @@ function BurgerConstructor() {
       }
     },
   });
-
-  // Получаем id всех ингредиентов, находящихся в конструкторе
-  const ingredientsId = useMemo(
-    () => getIngredientsId(ingredients.bun, ingredients.fillings),
-    [ingredients.bun, ingredients.fillings]
-  );
-
-  // Делаем запрос к api, получаем orderId и записываем его в глобальное состояние store.order.orderId
-  useEffect(() => {
-    if (!isEmptyBun) {
-      dispatch(setOrderId(ingredientsId));
-    }
-  }, [dispatch, ingredientsId, isEmptyBun]);
-
-  // Считаем общую стоимость заказа и записываем в глобальное состояние store.constructorBurger.total
-  useEffect(() => {
-    dispatch({
-      type: COUNT_TOTAL,
-      total: memoizedTotal,
-    });
-  }, [dispatch, memoizedTotal]);
 
   return (
     <section className={`mt-25 ${styles.constructor_section}`}>
