+++ src/components/burger-constructor/burger-constructor.module.css
@@ -1,5 +1,5 @@
 .constructor_section {
-  width: 50%;
+  width: 600px;
 }
 
 .constructor_container {
