+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,5 +1,5 @@
 .ingredients_section {
-  max-width: 50%;
+  max-width: 600px;
 }
 
 .heading {
