+++ src/services/actions/constructor.js
@@ -2,5 +2,4 @@ export const ADD_FILLING = "ADD_FILLING";
 export const ADD_BUN = "ADD_BUN";
 export const DELETE_BUN = "DELETE_BUN";
 export const DELETE_FILLING = "REMOVE_FILLING";
-export const COUNT_TOTAL = "COUNT_TOTAL";
 export const MOVE_FILLING = "MOVE_FILLING";
